requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
Django==2.2.5
django-filter==2.2.0
django-jquery==3.1.0
django-oauth-toolkit==1.2.0
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
nose==1.3.7
oauthlib==3.1.0
pipenv==2018.11.26
pylint==2.3.1
pytz==2019.2
requests==2.22.0
rest-condition==1.0.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.7.5
virtualenv-clone==0.5.3
wincertstore==0.2
wrapt==1.11.2
